+++ scratch.js
@@ -13,21 +13,19 @@ const wtf = require('./src/index');
 // let doc = readFile('BBDO');
 // console.log(doc.infoboxes(0).data);
 
-// wtf.fetch('Taylor%20Swift', 'en', {
-//   'Api-User-Agent': 'obfuscated@gmail.com'
-// }).then(docList => {
-//   console.warn(docList);
+// wtf.fetch('Горбатая_гора', 'ru', function(err, doc) {
+//   console.log(doc.sections('Сюжет').sentences().map((s) => s.text()));
 // });
 
-// let doc = wtf('{{some_template|link=[https://google.com google]}}');
-let str = `[[File:Volkswagen W12.jpg|thumb|upright|[[Volkswagen Group]] W12 engine from the [[Volkswagen Phaeton|Volkswagen Phaeton W12]]]]`;
-let doc = wtf(str);
-console.log(doc.images(0).json());
-// console.log(wtf.version);
 
-// console.log(doc.templates(0).rounds[0]);
-// console.log(doc.tables(0).keyValue());
-// console.log(doc.images(0).json());
-// console.log(wtf(str).sections('roster').templates('mlbplayer'));
-// console.log(doc.text());
-// Alan Bean marriage template
+let str = `In its broadest sense, a pest is a [[Competition (biology)|competitor]] of humanity.<ref>[http://www.merriam-webster.com/dictionary/pest Merriam-Webster dictionary], accessed 22 August 2012.</ref><ref>{{cite web |title=Pest vermin |url=http://www.britannica.com/EBchecked/topic/453421/pest |publisher=Britannica |accessdate=24 August 2016}}</ref> `;
+console.log(wtf(str).json());
+
+
+// let str = `годы asf`;
+// str = '4';
+// // const hasWord = new RegExp('[a-z\u00C0-\u00FF][a-z\u00C0-\u00FF]', 'iu');
+// // let hasWord = new RegExp('[\\w\u0430-\u044f]$', 'ui');
+// // let hasWord = new RegExp('\p{L}', 'ui');
+// let hasWord = /[\wа-я]$/;
+// console.log(hasWord.test(str));